Use authorization to access the recipe
Fill in created by id when creating a new recipe
Check if user can manage the recipe before editing
Check if user can manage the recipe before deleting
Refactor recipe service
Refactor tests for recipe service

+++ src/RecipeWebApp/Pages/Recipes/View.cshtml.cs
@@ -1,3 +1,4 @@
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 using RecipeWebApp.Services;
@@ -9,13 +10,17 @@ namespace RecipeWebApp.Pages.Recipes
     public class ViewModel : PageModel
     {
         private readonly IRecipeService _service;
+        private readonly IAuthorizationService _authService;
         private readonly ILogger<ViewModel> _logger;
 
         public RecipeDetailViewModel Recipe { get; set; }
 
-        public ViewModel(IRecipeService service, ILogger<ViewModel> logger)
+        public ViewModel(IRecipeService service,
+            IAuthorizationService authService,
+            ILogger<ViewModel> logger)
         {
             _service = service;
+            _authService = authService;
             _logger = logger;
         }
 
@@ -42,6 +47,19 @@ namespace RecipeWebApp.Pages.Recipes
 
         public async Task<IActionResult> OnPostDeleteAsync(int id)
         {
+            var recipeDetail = await _service.GetRecipe(id);
+            if (recipeDetail is null)
+            {
+                _logger.LogWarning("Recipe not found: {id}", id);
+                return NotFound();
+            }
+
+            var authResult = await _authService.AuthorizeAsync(User, recipeDetail, "CanManageRecipe");
+            if (!authResult.Succeeded)
+            {
+                return new ForbidResult();
+            }
+
             try
             {
                 await _service.DeleteRecipe(id);
